backend/requirements.txt
# Core dependencies
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
pydantic==1.10.7
httpx==0.24.0
pandas==2.0.0  # For CSV processing

# Database
sqlalchemy==1.4.48  # Downgraded to be compatible with snowflake-sqlalchemy
psycopg2-binary==2.9.6

# Connectors
snowflake-connector-python==3.0.3
snowflake-sqlalchemy==1.4.7
# fabric-sdk==0.4.0  # Removed - package doesn't exist in PyPI
azure-identity==1.12.0  # Added for Microsoft Fabric authentication
azure-core==1.26.4
# fivetran-sdk==1.0.0
# dbt-core==1.4.5
dbtc==0.2.0  # dbt Cloud Python API Client
boto3==1.26.115

# Testing
pytest==7.3.1
pytest-cov==4.1.0 